Move type guessing function
