+++ redash/query_runner/query_results.py
@@ -1,16 +1,12 @@
 import json
 import logging
-import numbers
 import re
 import sqlite3
 
-from dateutil import parser
-
 from redash import models
 from redash.permissions import has_access, not_view_only
-from redash.query_runner import (TYPE_BOOLEAN, TYPE_DATETIME, TYPE_FLOAT,
-                                 TYPE_INTEGER, TYPE_STRING, BaseQueryRunner,
-                                 register)
+from redash.query_runner import (TYPE_STRING, BaseQueryRunner,
+                                 register, guess_type)
 from redash.utils import JSONEncoder
 
 logger = logging.getLogger(__name__)
@@ -18,28 +14,6 @@ logger = logging.getLogger(__name__)
 
 class PermissionError(Exception):
     pass
-
-
-def _guess_type(value):
-    if value == '' or value is None:
-        return TYPE_STRING
-
-    if isinstance(value, numbers.Integral):
-        return TYPE_INTEGER
-
-    if isinstance(value, float):
-        return TYPE_FLOAT
-
-    if unicode(value).lower() in ('true', 'false'):
-        return TYPE_BOOLEAN
-
-    try:
-        parser.parse(value)
-        return TYPE_DATETIME
-    except (ValueError, OverflowError):
-        pass
-
-    return TYPE_STRING
 
 
 def extract_query_ids(query):
@@ -140,7 +114,7 @@ class Results(BaseQueryRunner):
 
                 for i, row in enumerate(cursor):
                     for j, col in enumerate(row):
-                        guess = _guess_type(col)
+                        guess = guess_type(col)
 
                         if columns[j]['type'] is None:
                             columns[j]['type'] = guess
